redux 적용 후, 메모 테스트 완료 및 검색창 분리
